Split JsonLineTest.
